viet 2:34

+++ GuiLayer/tabService.cs
@@ -104,11 +104,18 @@ namespace GuiLayer
 
 
 
-                        /*                  // Tạo form mới và truyền thông tin vào
-                                          frmClientInformation infoForm = new frmClientInformation(this);
-                                          infoForm.SetClientInformation(id, name, idCard, phone, email, address, gender);
-                                          infoForm.SetClientInformation(id);
-                                          infoForm.ShowDialog();*/
+                        // Lấy thông tin từ dòng được chọn
+
+                        string name = selectedRow.Cells["tenDichVu"].Value.ToString();
+                        string price = selectedRow.Cells["donGia"].Value.ToString();
+                        string loaiDichVu = selectedRow.Cells["loaiDichVu"].Value.ToString();
+
+                        frmServiceInfor frmService = new frmServiceInfor(this);
+
+                        frmService.setServiceInformation(name, price, loaiDichVu);
+                        frmService.setServiceId(id);
+                        frmService.ShowDialog();
+
                     }
                 }
                 else
@@ -121,42 +128,87 @@ namespace GuiLayer
 
         private void dataGridView2_CellContentClick(object sender, DataGridViewCellEventArgs e)
         {
+
+        }
+
+        private void dataGridView2_CellContentClick_1(object sender, DataGridViewCellEventArgs e)
+        {
             if (e.RowIndex >= 0)
             {
-                if (dataGridView2.Columns[e.ColumnIndex].HeaderText == "Delete")
+                DataGridViewRow selectedRow = dataGridView2.Rows[e.RowIndex];
+                string id = selectedRow.Cells["idThietBi"].Value.ToString();
+                if (!string.IsNullOrEmpty(id))
                 {
-
-                    DialogResult result = MessageBox.Show("Are you sure you want to delete?", "Comfirm to delete", MessageBoxButtons.YesNo, MessageBoxIcon.Question);
-
-                    if (result == DialogResult.Yes)
+                    if (dataGridView2.Columns[e.ColumnIndex].HeaderText == "Delete")
                     {
 
-                        DataGridViewRow selectedRow = dataGridView2.Rows[e.RowIndex];
-                        string id = selectedRow.Cells["idThietBi"].Value.ToString();
-                        int idAsInt = int.Parse(id);
+                        DialogResult result = MessageBox.Show("Are you sure you want to delete?", "Comfirm to delete", MessageBoxButtons.YesNo, MessageBoxIcon.Question);
 
-                        classThietBi thietBi = new classThietBi();
-                        thietBi.idThietBi = idAsInt;
-                        busThietBi.deleteThietBi(thietBi);
+                        if (result == DialogResult.Yes)
+                        {
 
-                        dataGridView2.Rows.RemoveAt(e.RowIndex);
-                        MessageBox.Show("Delete successfull");
+
+                            int idAsInt = int.Parse(id);
+
+                            classThietBi thietBi = new classThietBi();
+                            thietBi.idThietBi = idAsInt;
+                            busThietBi.deleteThietBi(thietBi);
+
+                            dataGridView2.Rows.RemoveAt(e.RowIndex);
+                            MessageBox.Show("Delete successfull");
+                        }
+
                     }
+                    else if (dataGridView2.Columns[e.ColumnIndex].HeaderText == "Edit")
+                    {
+                        string name = selectedRow.Cells["tenThietBi"].Value.ToString();
+                        string price = selectedRow.Cells["donGiaThietBI"].Value.ToString();
 
+
+                        frmFacilitiInfor facilitiInfor = new frmFacilitiInfor(this);
+
+                        facilitiInfor.setFacilitiInformation(name, price);
+                        facilitiInfor.setFacilitiID(id);
+                        facilitiInfor.ShowDialog();
+                    }
                 }
-                else if (dataGridView2.Columns[e.ColumnIndex].HeaderText == "Edit")
+            }
+        }
+
+        private void txtSearch_TextChanged_1(object sender, EventArgs e)
+        {
+            string search = txtSearch.Text.Trim();
+            classThietBi classThietBi = new classThietBi();
+            classDichVu classDichVu = new classDichVu();
+            foreach (Control control in panel1.Controls)
+            {
+                if (control is RadioButton radioButton)
                 {
-                    // Lấy thông tin từ dòng được chọn
-                    DataGridViewRow selectedRow = dataGridView1.Rows[e.RowIndex];
-                    string id = selectedRow.Cells["idKhachHang"].Value.ToString();
-                    string name = selectedRow.Cells["hoTen"].Value.ToString();
-                    string idCard = selectedRow.Cells["soCCCD"].Value.ToString();
-                    string phone = selectedRow.Cells["dienThoai"].Value.ToString();
-                    string email = selectedRow.Cells["email"].Value.ToString();
-                    string address = selectedRow.Cells["diaChi"].Value.ToString();
-                    string gender = selectedRow.Cells["gioiTinh"].Value.ToString();
+
+                    if (radioButton != null && radioButton.Checked)
+                    {
+                        string Option = radioButton.Text;
+                        if (Option == "Facilities")
+                        {
+                            classThietBi.tenThietBi = search;
+                            
+                            dataGridView2.DataSource = busThietBi.SearchThietBi(classThietBi);
+                        }
+                        else if (Option == "Services")
+                        {
+                            classDichVu.tenDichVu = search;
+                            busDichVu.SearchDichVu(classDichVu);
+                            dataGridView1.DataSource = busDichVu.SearchDichVu(classDichVu);
+                        }
+                        else
+                        {
+                            classDichVu.tenDichVu = search;
+                            busDichVu.SearchDichVu(classDichVu);
+                            dataGridView2.DataSource = busDichVu.SearchDichVu(classDichVu);
+                        }
 
 
+                    }
                 }
             }
         }
